Made parallel fitness functions optional
I built this code to use parallelizable
fitness functions (objective functions
which take a list of individuals rather
than a single one) from the beginning out
of experience. As a released library
though, it makes it hard to understand
for novices. I switched the parallel
evaluation optional and off by default.