Removed logistic duplicate
The previous implementation of `#logistic` was
just another form of `#sigmoid`, initially introduced
for performance testing and then forgotten.
Problem is, `exp(x)` is `Fload::INFINITY` for high
enough `x`, then `exp/exp` would return the
not-so-obvious value of `Float::NAN` as the network
activation.
Since it adds nothing to the current implementation,
I just removed it and aliased `#sigmoid`.